+++ autovpn.py
@@ -6,8 +6,12 @@ import toml
 
 def parse_config_file(path):
     """Parse TOML config file and return dictionary"""
-    with open(path, 'r') as f:
-        return toml.loads(f.read())
+    try:
+        with open(path, 'r') as f:
+            return toml.loads(f.read())
+    except:
+        open(path,'a').close() 
+        return {}
 
 def get_config_dict():
     return parse_config_file(f'{os.path.expanduser("~")}/.autovpnconfig')
@@ -20,7 +24,11 @@ def cli():
 def connect():
     """Connect to VPN"""
     settings = get_config_dict()
-    with Popen(['/opt/cisco/anyconnect/bin/vpn','connect',settings['server'],'-s'],stdin=PIPE,stdout=PIPE,text=True) as proc:
+    if settings == {}:
+        click.echo('autovpn has not been configured\nPlease configure using `autovpn configure`')
+        return
+    click.echo('Connecting to VPN...')
+    with Popen([settings['vpnpath'],'connect',settings['server'],'-s'],stdin=PIPE,stdout=PIPE,text=True) as proc:
         (stdout,stderr) = proc.communicate(f'{settings["username"]}\n{settings["password"]}\ny')
     output = stdout.split("\n")
     output = list(filter(None, output))
@@ -30,7 +38,11 @@ def connect():
 @cli.command()
 def state():
     """Show VPN state"""
-    with Popen(['/opt/cisco/anyconnect/bin/vpn','state'],stdout=PIPE,text=True) as proc:
+    settings = get_config_dict()
+    if settings == {}:
+        click.echo('autovpn has not been configured\nPlease configure using `autovpn configure`')
+        return
+    with Popen([settings['vpnpath'],'state'],stdout=PIPE,text=True) as proc:
         stdout = proc.stdout.read()
     output = stdout.split("\n")
     output = list(filter(None, output))
@@ -40,7 +52,11 @@ def state():
 @cli.command()
 def disconnect():
     """Disconnect from VPN"""
-    with Popen(['/opt/cisco/anyconnect/bin/vpn','disconnect'],stdout=PIPE,text=True) as proc:
+    settings = get_config_dict()
+    if settings == {}:
+        click.echo('autovpn has not been configured\nPlease configure using `autovpn configure`')
+        return
+    with Popen([settings['vpnpath'],'disconnect'],stdout=PIPE,text=True) as proc:
         stdout = proc.stdout.read()
     output = stdout.split("\n")
     output = list(filter(None, output))
@@ -51,13 +67,13 @@ def disconnect():
 def configure():
     """Configure VPN"""
     settings = get_config_dict()
-    settings['vpnpath'] = click.prompt(f'Enter VPN binary path', default=settings['vpnpath'], type=str)
-    settings['server'] = click.prompt(f'Enter VPN server', default=settings['server'], type=str)
-    settings['username'] = click.prompt(f'Enter VPN username', default=settings['username'], type=str)
-    settings['password'] = click.prompt(f'Enter VPN password', default=settings['password'], show_default = False, type=str)
+    settings['vpnpath'] = click.prompt(f'Enter VPN binary path', default=settings.get('vpnpath','/opt/cisco/anyconnect/bin/vpn'), type=str)
+    settings['server'] = click.prompt(f'Enter VPN server', default=settings.get('server','vpn.mobile.unibas.ch'), type=str)
+    settings['username'] = click.prompt(f'Enter VPN username', default=settings.get('username'), type=str)
+    settings['password'] = click.prompt(f'Enter VPN password {"[hidden]" if "password" in settings else ""}', default=settings.get('password'), hide_input=True, show_default = False, type=str)
     with open(f'{os.path.expanduser("~")}/.autovpnconfig', 'w') as f:
         f.write(toml.dumps(settings))
-    click.echo(f'Config file saved to {os.path.expanduser("~")}/.autovpnconfig')
+    click.echo(f'Configuration saved to {os.path.expanduser("~")}/.autovpnconfig')
 
 
 # TODO: Kill anyconnect UI